requirements.txt
together
python-dotenv
streamlit
google-cloud-firestore
langchain
langchain-community
langchain-text-splitters
pypdf
faiss-cpu
langchain-together
fastapi
uvicorn[standard]
python-multipart
requests
google-cloud-storage
openai
google-generativeai